core/engines/requirements.txt
# OCR/检测引擎模块最小依赖

# EasyOCR引擎依赖
easyocr>=1.7.0

# PaddleOCR引擎依赖
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Transformer引擎依赖
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# 通用依赖
Pillow>=10.0.0
numpy>=1.23.0
